requirements.txt
python-dotenv==1.0.0
gradio==3.50.2
langchain==0.0.334
tiktoken==0.5.1
faiss-cpu==1.7.4
torch==2.0.1
sentence-transformers==2.2.2
InstructorEmbedding==1.0.1
openai==1.2.3
loguru==0.7.2
textract==1.6.5
pandas==2.1.3
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
requests>=2.26.0,<3.0.0
